feat(landing): codex-style main-pane topic composer — the landing is the composer
Move topic/session creation out of the sidebar entirely. ChatHome now renders
TopicComposer: a message box centered in the main pane — type the first
prompt, hit enter, and the topic + session are created with that message.
Create/launch plumbing extracted to useTopicRoomCreate + lib/topic-create;
the sidebar TopicRoomCreatePanel, its state, and the
relay:open-topic-task-room event are deleted. All "+ task" buttons and the
palette action navigate to the landing via openTopicTaskRoom.
