+++ frontend/src/components/TopicSidebarShell.tsx
@@ -18,38 +18,27 @@ import {
 } from 'lucide-react';
 import type { WorkspaceSurface } from '../../../shared/workspace-surfaces.js';
 import {
-  buildWorkspaceTopicLaunchPreview,
   resolveTopicActiveContext,
   type WorkspaceTopic,
-  type WorkspaceTopicCreateInput,
-  type WorkspaceTopicLaunchIntent,
-  type WorkspaceTopicTemplateKind,
   type WorkspaceTopicSearchResult,
 } from '../../../shared/workspace-topics.js';
 import {
-  createWorkspaceTopicRoomAndMaybeLaunch,
   fetchHubNodes,
   fetchWorkspaceSurfaces,
   fetchWorkspaceTopics,
-  launchWorkspaceTopicRoom,
   restoreWorkspaceTopic,
   searchWorkspaceTopics,
-  type CreateSessionBody,
   sendSessionInput,
-  type WorkspaceTopicLaunchFailure,
-  type WorkspaceTopicRoomCreateResult,
 } from '../lib/api.js';
 import { deriveColor } from '../lib/colors.js';
-import { taskRefFromDraft } from '../lib/topic-task-ref.js';
+import { openTopicTaskRoom } from '../lib/topic-task-room.js';
 import type { SessionSummary } from '../lib/types.js';
 import { formatRelativeTimeCompact } from '../lib/utils.js';
 import { useSessionsStore } from '../lib/stores/sessions.js';
 import { useUiStore } from '../lib/stores/ui.js';
 import { useToastStore } from '../lib/stores/toasts.js';
 import { useIaWorkspacesQuery } from '../lib/hooks/use-ia-workspaces.js';
-import { useConfigStore } from '../lib/stores/config.js';
 import { durabilityDisabledReason } from '../lib/session-durability.js';
-import { resolveSessionByKey } from '../lib/session-keys.js';
 import {
   buildTopicNavModel,
   formatTaskRefLabel,
@@ -1279,455 +1268,6 @@ function topicEmptyStateText(input: {
   return `no topic matches for “${input.searchQuery.trim()}”`;
 }
 
-type TopicRoomDraft = {
-  title: string;
-  prompt: string;
-  taskRef: string;
-  providerId: string;
-  agentId: string;
-  nodeId: string;
-  repoPath: string;
-  worktreePath: string;
-  cwd: string;
-  templateKind: WorkspaceTopicTemplateKind;
-};
-
-const TOPIC_ROOM_DRAFT_EMPTY: TopicRoomDraft = {
-  title: '',
-  prompt: '',
-  taskRef: '',
-  providerId: '',
-  agentId: '',
-  nodeId: '',
-  repoPath: '',
-  worktreePath: '',
-  cwd: '',
-  templateKind: 'agent-task',
-};
-
-const TOPIC_ROOM_TEMPLATE_OPTIONS: Array<{
-  value: WorkspaceTopicTemplateKind;
-  label: string;
-}> = [
-  { value: 'agent-task', label: 'agent task' },
-  { value: 'terminal-task', label: 'terminal task' },
-  { value: 'note', label: 'note / room only' },
-];
-
-const FALLBACK_PROVIDER_IDS = ['claude', 'codex', 'opencode', 'hermes'];
-
-function compactString(value: string | null | undefined): string | undefined {
-  const trimmed = value?.trim();
-  return trimmed ? trimmed : undefined;
-}
-
-function uniqueStrings(values: Array<string | null | undefined>): string[] {
-  return Array.from(
-    new Set(values.map((value) => compactString(value)).filter(Boolean))
-  ) as string[];
-}
-
-/**
- * #1058: codex-style topic creation — the first message doubles as the room
- * title unless the operator overrides it in the advanced section.
- */
-function deriveTopicTitleFromPrompt(prompt: string): string {
-  const firstLine = prompt.trim().split('\n')[0] ?? '';
-  const collapsed = firstLine.replace(/\s+/g, ' ').trim();
-  if (!collapsed) return '';
-  // Code-point-aware truncation so a trailing emoji is not split in half.
-  const points = Array.from(collapsed);
-  return points.length > 60 ? `${points.slice(0, 59).join('')}…` : collapsed;
-}
-
-/** Title used for create/launch: explicit override wins, else the message. */
-function effectiveDraftTitle(draft: Pick<TopicRoomDraft, 'title' | 'prompt'>) {
-  return draft.title.trim() || deriveTopicTitleFromPrompt(draft.prompt);
-}
-
-function launchSubmitLabel(input: {
-  submittingIntent?: WorkspaceTopicLaunchIntent | null | undefined;
-  launchDisabled: boolean;
-  launchFailure?: WorkspaceTopicLaunchFailure | null | undefined;
-}): string {
-  if (input.submittingIntent === 'create-and-launch') return 'launching…';
-  if (input.launchDisabled) return 'note is create-only';
-  if (!input.launchFailure) return 'start';
-  return input.launchFailure.stage === 'session'
-    ? 'retry launch'
-    : 'retry create + launch';
-}
-
-function launchTypeForTemplate(
-  templateKind: WorkspaceTopicTemplateKind
-): CreateSessionBody['type'] | null {
-  if (templateKind === 'terminal-task') return 'terminal';
-  if (templateKind === 'agent-task') return 'agent';
-  return null;
-}
-
-function buildTopicRoomLaunchBody(
-  create: WorkspaceTopicCreateInput,
-  templateKind: WorkspaceTopicTemplateKind
-): Omit<CreateSessionBody, 'workspaceTopicId' | 'workContextId'> | null {
-  const type = launchTypeForTemplate(templateKind);
-  if (!type) return null;
-  const routing = create.routingDefaults ?? {};
-  return {
-    type,
-    mode: 'pty',
-    ...(type === 'agent' && routing.providerId
-      ? { agent: routing.providerId }
-      : {}),
-    ...(routing.nodeId ? { nodeId: routing.nodeId } : {}),
-    ...(routing.repoPath ? { repoPath: routing.repoPath } : {}),
-    ...(routing.worktreePath ? { worktreePath: routing.worktreePath } : {}),
-    ...(routing.cwd ? { cwd: routing.cwd } : {}),
-    controlMode: type === 'agent' ? 'agent-driven' : 'human-driven',
-  };
-}
-
-function buildTopicRoomCreateInput(input: {
-  draft: TopicRoomDraft;
-  workspaceId: string | null;
-  defaultProviderId: string;
-  defaultNodeId?: string | undefined;
-  defaultRepoPath?: string | undefined;
-  defaultWorktreePath?: string | undefined;
-  defaultCwd?: string | undefined;
-  taskRef: ReturnType<typeof taskRefFromDraft>;
-}): WorkspaceTopicCreateInput {
-  const providerId =
-    compactString(input.draft.providerId) ?? input.defaultProviderId;
-  const agentId = compactString(input.draft.agentId);
-  const nodeId = compactString(input.draft.nodeId) ?? input.defaultNodeId;
-  const repoPath = compactString(input.draft.repoPath) ?? input.defaultRepoPath;
-  const worktreePath =
-    compactString(input.draft.worktreePath) ?? input.defaultWorktreePath;
-  const cwd = compactString(input.draft.cwd) ?? input.defaultCwd;
-  const prompt = input.draft.prompt.trim();
-
-  return {
-    workspaceId: input.workspaceId ?? 'workspace:local',
-    title: input.draft.title.trim() || 'Untitled task room',
-    ...(prompt ? { description: prompt.slice(0, 240) } : {}),
-    promptDefaults: {
-      ...(prompt ? { starterPrompt: prompt } : {}),
-    },
-    routingDefaults: {
-      ...(providerId ? { providerId } : {}),
-      ...(agentId ? { agentId } : {}),
-      ...(nodeId ? { nodeId } : {}),
-      ...(repoPath ? { repoPath } : {}),
-      ...(worktreePath ? { worktreePath } : {}),
-      ...(cwd ? { cwd } : {}),
-    },
-    linkedRefs: {
-      ...(input.taskRef ? { taskRefs: [input.taskRef] } : {}),
-    },
-  };
-}
-
-interface TopicRoomCreatePanelProps {
-  open: boolean;
-  draft: TopicRoomDraft;
-  previewCreate: WorkspaceTopicCreateInput;
-  providerOptions: string[];
-  nodes: TopicNavNode[];
-  nodeOptions: Array<{ value: string; label: string }>;
-  repoPathOptions: string[];
-  worktreePathOptions: string[];
-  cwdOptions: string[];
-  launchFailure?: WorkspaceTopicLaunchFailure | null | undefined;
-  submittingIntent?: WorkspaceTopicLaunchIntent | null | undefined;
-  onDraftChange: (patch: Partial<TopicRoomDraft>) => void;
-  onSubmit: (intent: WorkspaceTopicLaunchIntent) => void;
-  onCancel: () => void;
-}
-
-function TopicRoomCreatePanel({
-  open,
-  draft,
-  previewCreate,
-  providerOptions,
-  nodes,
-  nodeOptions,
-  repoPathOptions,
-  worktreePathOptions,
-  cwdOptions,
-  launchFailure,
-  submittingIntent,
-  onDraftChange,
-  onSubmit,
-  onCancel,
-}: TopicRoomCreatePanelProps) {
-  const preview = useMemo(
-    () =>
-      buildWorkspaceTopicLaunchPreview({
-        create: previewCreate,
-        intent:
-          draft.templateKind === 'note' ? 'create-only' : 'create-and-launch',
-        templateKind: draft.templateKind,
-        launchOverrides: {
-          type: launchTypeForTemplate(draft.templateKind) ?? 'agent',
-          mode: 'pty',
-          agent: previewCreate.routingDefaults?.providerId,
-          nodeId: previewCreate.routingDefaults?.nodeId,
-          repoPath: previewCreate.routingDefaults?.repoPath,
-          worktreePath: previewCreate.routingDefaults?.worktreePath,
-          cwd: previewCreate.routingDefaults?.cwd,
-        },
-      }),
-    [draft.templateKind, previewCreate]
-  );
-  const [advancedOpen, setAdvancedOpen] = useState(false);
-  if (!open) return null;
-  const launchDisabled = draft.templateKind === 'note';
-  const disabled = !effectiveDraftTitle(draft) || Boolean(submittingIntent);
-  // #1103: never render a raw node id — resolve through the roster or fall
-  // back to a generic label.
-  const routedNodeId = previewCreate.routingDefaults?.nodeId;
-  const friendlyNodeLabel = routedNodeId
-    ? nodes.find((node) => node.nodeId === routedNodeId)?.displayName ||
-      'remote node'
-    : preview.nodeLabel;
-  return (
-    <form
-      className="topic-create-panel"
-      aria-label="new topic"
-      onSubmit={(event: FormEvent) => {
-        event.preventDefault();
-        if (!disabled && !launchDisabled) onSubmit('create-and-launch');
-      }}
-    >
-      <div className="topic-create-panel__title">new topic</div>
-      <textarea
-        value={draft.prompt}
-        onChange={(event) => onDraftChange({ prompt: event.target.value })}
-        onKeyDown={(event) => {
-          if (
-            event.key === 'Enter' &&
-            !event.shiftKey &&
-            !event.nativeEvent.isComposing &&
-            !disabled &&
-            !launchDisabled
-          ) {
-            event.preventDefault();
-            onSubmit('create-and-launch');
-          }
-        }}
-        placeholder="what should the agent do?"
-        rows={3}
-        aria-label="first message"
-        autoFocus
-      />
-      <div className="topic-create-panel__context">
-        {preview.providerLabel} · {friendlyNodeLabel} · {preview.cwdLabel}
-      </div>
-      <button
-        type="button"
-        className="topic-create-panel__advanced-toggle"
-        aria-expanded={advancedOpen}
-        onClick={() => setAdvancedOpen((prev) => !prev)}
-      >
-        <span aria-hidden="true">{advancedOpen ? '▾ ' : '▸ '}</span>
-        advanced
-      </button>
-      {advancedOpen ? (
-        <>
-          <label>
-            <span>title</span>
-            <input
-              value={draft.title}
-              onChange={(event) => onDraftChange({ title: event.target.value })}
-              placeholder={
-                deriveTopicTitleFromPrompt(draft.prompt) || 'auto from message'
-              }
-            />
-          </label>
-          <label>
-            <span>task ref</span>
-            <input
-              value={draft.taskRef}
-              onChange={(event) =>
-                onDraftChange({ taskRef: event.target.value })
-              }
-              placeholder="github issue number or URL"
-            />
-          </label>
-          <label>
-            <span>provider</span>
-            <input
-              list="topic-room-provider-options"
-              value={draft.providerId}
-              onChange={(event) =>
-                onDraftChange({ providerId: event.target.value })
-              }
-              placeholder={
-                previewCreate.routingDefaults?.providerId ?? 'default provider'
-              }
-            />
-            <datalist id="topic-room-provider-options">
-              {providerOptions.map((providerId) => (
-                <option key={providerId} value={providerId} />
-              ))}
-            </datalist>
-          </label>
-          <label>
-            <span>agent id</span>
-            <input
-              value={draft.agentId}
-              onChange={(event) =>
-                onDraftChange({ agentId: event.target.value })
-              }
-              placeholder="optional agent identity"
-            />
-          </label>
-          <label>
-            <span>template kind</span>
-            <select
-              value={draft.templateKind}
-              onChange={(event) =>
-                onDraftChange({
-                  templateKind: event.target
-                    .value as WorkspaceTopicTemplateKind,
-                })
-              }
-            >
-              {TOPIC_ROOM_TEMPLATE_OPTIONS.map((option) => (
-                <option key={option.value} value={option.value}>
-                  {option.label}
-                </option>
-              ))}
-            </select>
-          </label>
-          <label>
-            <span>node</span>
-            <input
-              list="topic-room-node-options"
-              value={draft.nodeId}
-              onChange={(event) =>
-                onDraftChange({ nodeId: event.target.value })
-              }
-              placeholder={
-                previewCreate.routingDefaults?.nodeId ?? 'local/default node'
-              }
-            />
-            <datalist id="topic-room-node-options">
-              {nodeOptions.map((node) => (
-                <option
-                  key={node.value}
-                  value={node.value}
-                  label={node.label}
-                />
-              ))}
-            </datalist>
-          </label>
-          <label>
-            <span>repo</span>
-            <input
-              list="topic-room-repo-options"
-              value={draft.repoPath}
-              onChange={(event) =>
-                onDraftChange({ repoPath: event.target.value })
-              }
-              placeholder={
-                previewCreate.routingDefaults?.repoPath ?? 'default repo'
-              }
-            />
-            <datalist id="topic-room-repo-options">
-              {repoPathOptions.map((repoPath) => (
-                <option key={repoPath} value={repoPath} />
-              ))}
-            </datalist>
-          </label>
-          <label>
-            <span>worktree</span>
-            <input
-              list="topic-room-worktree-options"
-              value={draft.worktreePath}
-              onChange={(event) =>
-                onDraftChange({ worktreePath: event.target.value })
-              }
-              placeholder={
-                previewCreate.routingDefaults?.worktreePath ??
-                'default worktree'
-              }
-            />
-            <datalist id="topic-room-worktree-options">
-              {worktreePathOptions.map((worktreePath) => (
-                <option key={worktreePath} value={worktreePath} />
-              ))}
-            </datalist>
-          </label>
-          <label>
-            <span>cwd</span>
-            <input
-              list="topic-room-cwd-options"
-              value={draft.cwd}
-              onChange={(event) => onDraftChange({ cwd: event.target.value })}
-              placeholder={previewCreate.routingDefaults?.cwd ?? 'default cwd'}
-            />
-            <datalist id="topic-room-cwd-options">
-              {cwdOptions.map((cwd) => (
-                <option key={cwd} value={cwd} />
-              ))}
-            </datalist>
-          </label>
-          <div className="topic-create-preview" aria-label="launch preview">
-            <div>template: {preview.templateKind}</div>
-            <div>provider: {preview.providerLabel}</div>
-            <div>
-              agent:{' '}
-              {previewCreate.routingDefaults?.agentId ??
-                previewCreate.routingDefaults?.providerId ??
-                'default agent'}
-            </div>
-            <div>mode: {preview.modeLabel}</div>
-            <div>node: {friendlyNodeLabel}</div>
-            <div>cwd: {preview.cwdLabel}</div>
-            <div>prompt: {preview.promptSources.join(', ')}</div>
-            <div>tasks: {preview.taskRefs.join(', ')}</div>
-            <div>side effects: {preview.sideEffects.join(' · ')}</div>
-          </div>
-        </>
-      ) : null}
-      {launchFailure ? (
-        <div className="topic-create-failure" role="alert">
-          {launchFailure.stage === 'session'
-            ? 'launch failed after room creation'
-            : 'room creation failed'}{' '}
-          ({launchFailure.stage}): {launchFailure.message}
-        </div>
-      ) : null}
-      <div className="topic-create-panel__actions">
-        <button
-          type="button"
-          onClick={onCancel}
-          disabled={Boolean(submittingIntent)}
-        >
-          cancel
-        </button>
-        {advancedOpen || launchDisabled ? (
-          <button
-            type="button"
-            disabled={disabled}
-            onClick={() => onSubmit('create-only')}
-          >
-            {submittingIntent === 'create-only' ? 'creating…' : 'create only'}
-          </button>
-        ) : null}
-        <button type="submit" disabled={disabled || launchDisabled}>
-          {launchSubmitLabel({
-            submittingIntent,
-            launchDisabled,
-            launchFailure,
-          })}
-        </button>
-      </div>
-    </form>
-  );
-}
 
 /** A workspace bucket of mobile topic rows, kept in attention order. */
 interface MobileTopicGroup {
@@ -2030,7 +1570,6 @@ export function TopicSidebarView({
   onSearchClear,
   onSelectSession,
   onSendInput = sendSessionInput,
-  createPanel,
   onCreateTaskRoom,
   workspaces = EMPTY_WORKSPACES,
   nodes,
@@ -2071,7 +1610,6 @@ export function TopicSidebarView({
   onSearchClear?: (() => void) | undefined;
   onSelectSession?: ((id: string) => void) | undefined;
   onSendInput?: TopicSendInput | undefined;
-  createPanel?: ReactNode;
   onCreateTaskRoom?: (() => void) | undefined;
 }) {
   const model = useMemo(
@@ -2247,7 +1785,6 @@ export function TopicSidebarView({
             : undefined
         }
       />
-      {createPanel}
       <TopicSearchPanel
         model={model}
         searchQuery={searchQuery}
@@ -2298,30 +1835,11 @@ export function TopicSidebarShell({
 }) {
   const queryClient = useQueryClient();
   const sessions = useSessionsStore((s) => s.sessions);
-  const activeSessionId = useSessionsStore((s) => s.activeSessionId);
-  const setActiveSessionId = useSessionsStore((s) => s.setActiveSessionId);
-  const activeRepoPath = useUiStore((s) => s.activeRepoPath);
   const activeWorkspaceId = useUiStore((s) => s.activeWorkspaceId);
-  const defaultAgent = useConfigStore((s) => s.defaultAgent);
-  const frameworks = useConfigStore((s) => s.frameworks);
-  const activeSession = useMemo(
-    () => resolveSessionByKey(sessions, activeSessionId),
-    [activeSessionId, sessions]
-  );
   const [searchQuery, setSearchQuery] = useState('');
   const [searchScope, setSearchScope] = useState<'all' | 'workspace'>('all');
   const scopedWorkspaceId =
     searchScope === 'workspace' ? activeWorkspaceId : null;
-  const [createOpen, setCreateOpen] = useState(false);
-  const [createDraft, setCreateDraft] = useState<TopicRoomDraft>(
-    TOPIC_ROOM_DRAFT_EMPTY
-  );
-  const [submittingIntent, setSubmittingIntent] =
-    useState<WorkspaceTopicLaunchIntent | null>(null);
-  const [launchFailure, setLaunchFailure] =
-    useState<WorkspaceTopicLaunchFailure | null>(null);
-  const [createdRoom, setCreatedRoom] =
-    useState<WorkspaceTopicRoomCreateResult | null>(null);
   const normalizedSearchQuery = searchQuery.trim();
   const [showArchived, setShowArchived] = useState(false);
   const topicsQuery = useQuery({
@@ -2377,15 +1895,6 @@ export function TopicSidebarShell({
     queryFn: fetchHubNodes,
     staleTime: 60_000,
   });
-  useEffect(() => {
-    const openCreate = () => {
-      setCreateOpen(true);
-      setLaunchFailure(null);
-    };
-    window.addEventListener('relay:open-topic-task-room', openCreate);
-    return () =>
-      window.removeEventListener('relay:open-topic-task-room', openCreate);
-  }, []);
   const searchActive = normalizedSearchQuery.length > 0;
   const searchData = topicSearchQuery.data;
   const searchResults = useMemo(
@@ -2413,196 +1922,6 @@ export function TopicSidebarShell({
       })),
     [nodesQuery.data]
   );
-  const effectiveCreateTitle = effectiveDraftTitle(createDraft);
-  const taskRef = taskRefFromDraft(createDraft.taskRef, effectiveCreateTitle);
-  const defaultRepoPath =
-    activeSession?.repoPath ?? activeRepoPath ?? undefined;
-  const defaultWorktreePath = activeSession?.worktreePath ?? undefined;
-  const defaultCwd =
-    activeSession?.cwd ?? defaultWorktreePath ?? defaultRepoPath ?? undefined;
-  const providerOptions = useMemo(
-    () =>
-      uniqueStrings([
-        defaultAgent,
-        ...frameworks.map((framework) => framework.id),
-        ...FALLBACK_PROVIDER_IDS,
-      ]),
-    [defaultAgent, frameworks]
-  );
-  const nodeOptions = useMemo(
-    () =>
-      (nodesQuery.data ?? []).map((node) => ({
-        value: node.nodeId,
-        label: node.displayName
-          ? `${node.displayName} · ${node.status}`
-          : node.status,
-      })),
-    [nodesQuery.data]
-  );
-  const repoPathOptions = useMemo(
-    () =>
-      uniqueStrings([
-        defaultRepoPath,
-        ...sessions.map((session) => session.repoPath),
-      ]),
-    [defaultRepoPath, sessions]
-  );
-  const worktreePathOptions = useMemo(
-    () =>
-      uniqueStrings([
-        defaultWorktreePath,
-        ...sessions.map((session) => session.worktreePath),
-      ]),
-    [defaultWorktreePath, sessions]
-  );
-  const cwdOptions = useMemo(
-    () =>
-      uniqueStrings([defaultCwd, ...sessions.map((session) => session.cwd)]),
-    [defaultCwd, sessions]
-  );
-  const previewCreate = useMemo<WorkspaceTopicCreateInput>(
-    () =>
-      buildTopicRoomCreateInput({
-        draft: { ...createDraft, title: effectiveCreateTitle },
-        workspaceId: activeWorkspaceId,
-        defaultProviderId: defaultAgent,
-        defaultNodeId: activeSession?.nodeId,
-        defaultRepoPath,
-        defaultWorktreePath,
-        defaultCwd,
-        taskRef,
-      }),
-    [
-      activeSession?.nodeId,
-      activeWorkspaceId,
-      createDraft,
-      defaultAgent,
-      defaultCwd,
-      defaultRepoPath,
-      defaultWorktreePath,
-      effectiveCreateTitle,
-      taskRef,
-    ]
-  );
-
-  const handleCreateSubmit = useCallback(
-    async (intent: WorkspaceTopicLaunchIntent) => {
-      if (!effectiveCreateTitle) return;
-      const launch = buildTopicRoomLaunchBody(
-        previewCreate,
-        createDraft.templateKind
-      );
-      const submitIntent =
-        intent === 'create-and-launch' && launch
-          ? 'create-and-launch'
-          : 'create-only';
-      setSubmittingIntent(submitIntent);
-      setLaunchFailure(null);
-      try {
-        if (submitIntent === 'create-and-launch' && createdRoom && launch) {
-          const result = await launchWorkspaceTopicRoom({
-            room: createdRoom,
-            launch,
-          });
-          if (result.status === 'launch_failed') {
-            setLaunchFailure(result.failure);
-            return;
-          }
-          await useSessionsStore.getState().refreshAll();
-          setActiveSessionId(result.session.id);
-          onSelectSession?.(result.session.id);
-          setCreateOpen(false);
-          setCreatedRoom(null);
-          setCreateDraft(TOPIC_ROOM_DRAFT_EMPTY);
-          return;
-        }
-        const result = await createWorkspaceTopicRoomAndMaybeLaunch({
-          room: {
-            topic: previewCreate,
-            ...(taskRef ? { taskRef } : {}),
-          },
-          ...(submitIntent === 'create-and-launch' && launch
-            ? {
-                launch,
-              }
-            : {}),
-        });
-        await Promise.all([
-          queryClient.invalidateQueries({ queryKey: ['workspace-topics'] }),
-          queryClient.invalidateQueries({ queryKey: ['workspace-surfaces'] }),
-        ]);
-        if (result.status === 'launch_failed') {
-          setLaunchFailure(result.failure);
-          setCreatedRoom({
-            topic: result.topic,
-            workContext: result.workContext,
-          });
-          return;
-        }
-        if (result.status === 'launched') {
-          await useSessionsStore.getState().refreshAll();
-          setActiveSessionId(result.session.id);
-          onSelectSession?.(result.session.id);
-        }
-        setCreateOpen(false);
-        setCreatedRoom(null);
-        setCreateDraft(TOPIC_ROOM_DRAFT_EMPTY);
-      } catch (error) {
-        const failure = error as WorkspaceTopicLaunchFailure;
-        setLaunchFailure({
-          stage: failure.stage ?? 'topic',
-          message:
-            typeof failure.message === 'string'
-              ? failure.message
-              : error instanceof Error
-                ? error.message
-                : String(error),
-          retryable: failure.retryable ?? false,
-          ...(failure.code ? { code: failure.code } : {}),
-          ...(failure.status ? { status: failure.status } : {}),
-        });
-      } finally {
-        setSubmittingIntent(null);
-      }
-    },
-    [
-      effectiveCreateTitle,
-      createDraft.templateKind,
-      createdRoom,
-      onSelectSession,
-      previewCreate,
-      queryClient,
-      setActiveSessionId,
-      taskRef,
-    ]
-  );
-
-  const createPanel = (
-    <TopicRoomCreatePanel
-      open={createOpen}
-      draft={createDraft}
-      previewCreate={previewCreate}
-      providerOptions={providerOptions}
-      nodes={viewNodes}
-      nodeOptions={nodeOptions}
-      repoPathOptions={repoPathOptions}
-      worktreePathOptions={worktreePathOptions}
-      cwdOptions={cwdOptions}
-      launchFailure={launchFailure}
-      submittingIntent={submittingIntent}
-      onDraftChange={(patch) => {
-        setCreateDraft((current) => ({ ...current, ...patch }));
-        setCreatedRoom(null);
-        setLaunchFailure(null);
-      }}
-      onSubmit={(intent) => void handleCreateSubmit(intent)}
-      onCancel={() => {
-        setCreateOpen(false);
-        setLaunchFailure(null);
-      }}
-    />
-  );
-
   return (
     <TopicSidebarView
       topics={viewTopics}
@@ -2642,11 +1961,7 @@ export function TopicSidebarShell({
       onSearchRetry={() => void topicSearchQuery.refetch()}
       onSearchClear={() => setSearchQuery('')}
       onSelectSession={onSelectSession}
-      createPanel={createPanel}
-      onCreateTaskRoom={() => {
-        setCreateOpen(true);
-        setLaunchFailure(null);
-      }}
+      onCreateTaskRoom={openTopicTaskRoom}
     />
   );
 }
