+++ frontend/src/lib/topic-create.ts
@@ -1,0 +1,167 @@
+import type {
+  WorkspaceTopicCreateInput,
+  WorkspaceTopicLaunchIntent,
+  WorkspaceTopicTemplateKind,
+} from '../../../shared/workspace-topics.js';
+import type {
+  CreateSessionBody,
+  WorkspaceTopicLaunchFailure,
+} from './api.js';
+import type { taskRefFromDraft } from './topic-task-ref.js';
+
+/**
+ * #1058: pure draft/build helpers for codex-style topic creation, shared by
+ * the main-pane TopicComposer (primary path) and any future create surfaces.
+ * The first message doubles as the room title unless the operator overrides
+ * it in the advanced section.
+ */
+export type TopicRoomDraft = {
+  title: string;
+  prompt: string;
+  taskRef: string;
+  providerId: string;
+  agentId: string;
+  nodeId: string;
+  repoPath: string;
+  worktreePath: string;
+  cwd: string;
+  templateKind: WorkspaceTopicTemplateKind;
+};
+
+export const TOPIC_ROOM_DRAFT_EMPTY: TopicRoomDraft = {
+  title: '',
+  prompt: '',
+  taskRef: '',
+  providerId: '',
+  agentId: '',
+  nodeId: '',
+  repoPath: '',
+  worktreePath: '',
+  cwd: '',
+  templateKind: 'agent-task',
+};
+
+export const TOPIC_ROOM_TEMPLATE_OPTIONS: Array<{
+  value: WorkspaceTopicTemplateKind;
+  label: string;
+}> = [
+  { value: 'agent-task', label: 'agent task' },
+  { value: 'terminal-task', label: 'terminal task' },
+  { value: 'note', label: 'note / room only' },
+];
+
+export const FALLBACK_PROVIDER_IDS = ['claude', 'codex', 'opencode', 'hermes'];
+
+export function compactString(
+  value: string | null | undefined
+): string | undefined {
+  const trimmed = value?.trim();
+  return trimmed ? trimmed : undefined;
+}
+
+export function uniqueStrings(
+  values: Array<string | null | undefined>
+): string[] {
+  return Array.from(
+    new Set(values.map((value) => compactString(value)).filter(Boolean))
+  ) as string[];
+}
+
+export function deriveTopicTitleFromPrompt(prompt: string): string {
+  const firstLine = prompt.trim().split('\n')[0] ?? '';
+  const collapsed = firstLine.replace(/\s+/g, ' ').trim();
+  if (!collapsed) return '';
+  // Code-point-aware truncation so a trailing emoji is not split in half.
+  const points = Array.from(collapsed);
+  return points.length > 60 ? `${points.slice(0, 59).join('')}…` : collapsed;
+}
+
+/** Title used for create/launch: explicit override wins, else the message. */
+export function effectiveDraftTitle(
+  draft: Pick<TopicRoomDraft, 'title' | 'prompt'>
+) {
+  return draft.title.trim() || deriveTopicTitleFromPrompt(draft.prompt);
+}
+
+export function launchSubmitLabel(input: {
+  submittingIntent?: WorkspaceTopicLaunchIntent | null | undefined;
+  launchDisabled: boolean;
+  launchFailure?: WorkspaceTopicLaunchFailure | null | undefined;
+}): string {
+  if (input.submittingIntent === 'create-and-launch') return 'launching…';
+  if (input.launchDisabled) return 'note is create-only';
+  if (!input.launchFailure) return 'start';
+  return input.launchFailure.stage === 'session'
+    ? 'retry launch'
+    : 'retry create + launch';
+}
+
+export function launchTypeForTemplate(
+  templateKind: WorkspaceTopicTemplateKind
+): CreateSessionBody['type'] | null {
+  if (templateKind === 'terminal-task') return 'terminal';
+  if (templateKind === 'agent-task') return 'agent';
+  return null;
+}
+
+export function buildTopicRoomLaunchBody(
+  create: WorkspaceTopicCreateInput,
+  templateKind: WorkspaceTopicTemplateKind
+): Omit<CreateSessionBody, 'workspaceTopicId' | 'workContextId'> | null {
+  const type = launchTypeForTemplate(templateKind);
+  if (!type) return null;
+  const routing = create.routingDefaults ?? {};
+  return {
+    type,
+    mode: 'pty',
+    ...(type === 'agent' && routing.providerId
+      ? { agent: routing.providerId }
+      : {}),
+    ...(routing.nodeId ? { nodeId: routing.nodeId } : {}),
+    ...(routing.repoPath ? { repoPath: routing.repoPath } : {}),
+    ...(routing.worktreePath ? { worktreePath: routing.worktreePath } : {}),
+    ...(routing.cwd ? { cwd: routing.cwd } : {}),
+    controlMode: type === 'agent' ? 'agent-driven' : 'human-driven',
+  };
+}
+
+export function buildTopicRoomCreateInput(input: {
+  draft: TopicRoomDraft;
+  workspaceId: string | null;
+  defaultProviderId: string;
+  defaultNodeId?: string | undefined;
+  defaultRepoPath?: string | undefined;
+  defaultWorktreePath?: string | undefined;
+  defaultCwd?: string | undefined;
+  taskRef: ReturnType<typeof taskRefFromDraft>;
+}): WorkspaceTopicCreateInput {
+  const providerId =
+    compactString(input.draft.providerId) ?? input.defaultProviderId;
+  const agentId = compactString(input.draft.agentId);
+  const nodeId = compactString(input.draft.nodeId) ?? input.defaultNodeId;
+  const repoPath = compactString(input.draft.repoPath) ?? input.defaultRepoPath;
+  const worktreePath =
+    compactString(input.draft.worktreePath) ?? input.defaultWorktreePath;
+  const cwd = compactString(input.draft.cwd) ?? input.defaultCwd;
+  const prompt = input.draft.prompt.trim();
+
+  return {
+    workspaceId: input.workspaceId ?? 'workspace:local',
+    title: input.draft.title.trim() || 'Untitled task room',
+    ...(prompt ? { description: prompt.slice(0, 240) } : {}),
+    promptDefaults: {
+      ...(prompt ? { starterPrompt: prompt } : {}),
+    },
+    routingDefaults: {
+      ...(providerId ? { providerId } : {}),
+      ...(agentId ? { agentId } : {}),
+      ...(nodeId ? { nodeId } : {}),
+      ...(repoPath ? { repoPath } : {}),
+      ...(worktreePath ? { worktreePath } : {}),
+      ...(cwd ? { cwd } : {}),
+    },
+    linkedRefs: {
+      ...(input.taskRef ? { taskRefs: [input.taskRef] } : {}),
+    },
+  };
+}
