+++ frontend/src/lib/topic-task-room.ts
@@ -1,19 +1,19 @@
+import { useSessionsStore } from './stores/sessions.js';
 import { useUiStore } from './stores/ui.js';
 
 /**
- * #1058: shared "start a topic" entry point used by both the work-cockpit
- * empty state and the chat-first landing home. The topic create panel lives
- * in TopicSidebarShell, which is unmounted while the sidebar is collapsed —
- * so un-collapse + open first, then dispatch on the next tick once the shell
- * has mounted and registered its listener (otherwise the event fires into
- * the void and the CTA is a no-op in collapsed mode).
+ * #1058: shared "start a topic" entry point. The composer IS the landing view
+ * (TopicComposer inside ChatHome), so starting a topic means navigating to the
+ * landing: clear the active session/repo/analytics selection and drop any
+ * forced cockpit so resolveAppViewMode returns 'chat'. The composer autofocuses
+ * its message box on mount. On mobile the sidebar drawer is closed so the main
+ * pane is actually visible.
  */
 export function openTopicTaskRoom(): void {
+  useSessionsStore.getState().setActiveSessionId(null);
   const ui = useUiStore.getState();
-  if (ui.sidebarCollapsed) ui.toggleSidebarCollapsed();
-  ui.openSidebar();
-  window.setTimeout(
-    () => window.dispatchEvent(new Event('relay:open-topic-task-room')),
-    0
-  );
+  ui.setActiveRepoPath(null);
+  ui.setAnalyticsView(null);
+  ui.setForceOrgCockpit(false);
+  ui.closeSidebar();
 }
