+++ test/topic-sidebar-shell.test.ts
@@ -851,22 +851,11 @@ describe('TopicSidebarView', () => {
     expect(container.textContent).toContain('no workspace topics yet');
   });
 
-  it('surfaces task-room creation entrypoints and preview panel', async () => {
+  it('routes the task-room creation entrypoint to the main-pane composer', async () => {
+    // #1058: creation lives in the main pane (TopicComposer), not a sidebar
+    // panel — the sidebar button only fires the navigation callback.
     const onCreateTaskRoom = vi.fn();
-    await renderView({
-      createPanel: React.createElement(
-        'form',
-        { className: 'topic-create-panel', 'aria-label': 'create task room' },
-        React.createElement(
-          'div',
-          { className: 'topic-create-preview' },
-          'provider: hermes'
-        ),
-        React.createElement('button', { type: 'button' }, 'create only'),
-        React.createElement('button', { type: 'button' }, 'create + launch')
-      ),
-      onCreateTaskRoom,
-    });
+    await renderView({ onCreateTaskRoom });
 
     const createButton = container.querySelector(
       '.topic-shell__create'
@@ -874,9 +863,7 @@ describe('TopicSidebarView', () => {
     await act(async () => createButton.click());
 
     expect(onCreateTaskRoom).toHaveBeenCalled();
-    expect(container.textContent).toContain('provider: hermes');
-    expect(container.textContent).toContain('create only');
-    expect(container.textContent).toContain('create + launch');
+    expect(container.querySelector('.topic-create-panel')).toBeNull();
   });
 
   it('renders a phone-first attention list sorted before routine topics', async () => {
